feat: add bria subscription reconnection
