+++ src/domain/primitives/index.types.d.ts
@@ -13,3 +13,16 @@ type Memo = string & { readonly brand: unique symbol }
 type XOR<T1, T2> =
   | (T1 & { [k in Exclude<keyof T2, keyof T1>]?: never })
   | (T2 & { [k in Exclude<keyof T1, keyof T2>]?: never })
+
+type JSONValue =
+  | string
+  | number
+  | bigint
+  | boolean
+  | Date
+  | JSONObject
+  | JSONArray
+  | null
+  | undefined
+type JSONArray = Array<JSONValue>
+type JSONObject = { [key: string]: JSONValue }
