+++ src/servers/event-handlers/bria.ts
@@ -2,11 +2,14 @@ import { Wallets } from "@app"
 
 import { CouldNotFindWalletFromOnChainAddressError } from "@domain/errors"
 
+import { baseLogger } from "@services/logger"
 import { BriaPayloadType } from "@services/bria"
 
-export const briaEventHandler = async (
-  event: BriaEvent,
-): Promise<true | ApplicationError> => {
+export const briaEventHandler = async (event: BriaEvent): Promise<true | DomainError> => {
+  baseLogger.info(
+    { sequence: event.sequence, type: event.payload.type },
+    "bria event handler",
+  )
   switch (event.payload.type) {
     case BriaPayloadType.UtxoDetected: {
       return utxoDetectedEventHandler({ event: event.payload })
@@ -24,7 +27,7 @@ export const utxoDetectedEventHandler = async ({
   event,
 }: {
   event: UtxoDetected
-}): Promise<true | ApplicationError> => {
+}): Promise<true | DomainError> => {
   const res = await Wallets.addPendingTransaction(event)
   if (res instanceof CouldNotFindWalletFromOnChainAddressError) {
     return true
@@ -37,7 +40,7 @@ export const utxoSettledEventHandler = async ({
   event,
 }: {
   event: UtxoSettled
-}): Promise<true | ApplicationError> => {
+}): Promise<true | DomainError> => {
   const res = await Wallets.addSettledTransaction(event)
   if (res instanceof CouldNotFindWalletFromOnChainAddressError) {
     return true
