+++ src/services/bria/index.ts
@@ -1,8 +1,5 @@
-import util from "util"
-
+import { Metadata, status } from "@grpc/grpc-js"
 import { Struct, Value, ListValue } from "google-protobuf/google/protobuf/struct_pb"
-
-import { credentials, Metadata, status } from "@grpc/grpc-js"
 
 import { getBriaConfig } from "@config"
 
@@ -13,73 +10,31 @@ import {
 } from "@domain/bitcoin/onchain"
 import { paymentAmountFromNumber, WalletCurrency } from "@domain/shared/primitives"
 
-import {
-  asyncRunInSpan,
-  SemanticAttributes,
-  recordExceptionInCurrentSpan,
-  wrapAsyncFunctionsToRunInSpan,
-} from "@services/tracing"
+import { baseLogger } from "@services/logger"
+import { wrapAsyncToRunInSpan, wrapAsyncFunctionsToRunInSpan } from "@services/tracing"
 
-import { BriaEventRepo } from "./repo"
-import { ListenerWrapper } from "./listener_wrapper"
-import { BriaServiceClient } from "./proto/bria_grpc_pb"
+import { GrpcStreamClient } from "@utils"
+
+import {
+  findAddressByExternalId,
+  getWalletBalanceSummary,
+  newAddress,
+  subscribeAll,
+} from "./grpc-client"
 import {
   SubscribeAllRequest,
   BriaEvent as RawBriaEvent,
   NewAddressRequest,
-  NewAddressResponse,
   GetWalletBalanceSummaryRequest,
-  GetWalletBalanceSummaryResponse,
   FindAddressByExternalIdRequest,
-  FindAddressByExternalIdResponse,
 } from "./proto/bria_pb"
-import {
-  EventAugmentationMissingError,
-  ExpectedAddressInfoMissingInEventError,
-  ExpectedPayoutBroadcastPayloadNotFoundError,
-  ExpectedPayoutCommittedPayloadNotFoundError,
-  ExpectedPayoutSettledPayloadNotFoundError,
-  ExpectedPayoutSubmittedPayloadNotFoundError,
-  ExpectedUtxoDetectedPayloadNotFoundError,
-  ExpectedUtxoSettledPayloadNotFoundError,
-  NoPayloadFoundError,
-  UnknownBriaEventError,
-  UnknownPayloadTypeReceivedError,
-} from "./errors"
-
-export { ListenerWrapper } from "./listener_wrapper"
+import { UnknownBriaEventError } from "./errors"
+export { BriaPayloadType } from "./event-handler"
+import { BriaEventRepo } from "./event-repository"
+import { eventDataHandler } from "./event-handler"
 
 const briaConfig = getBriaConfig()
-
-const bitcoinBridgeClient = new BriaServiceClient(
-  briaConfig.endpoint,
-  credentials.createInsecure(),
-)
-
-const newAddress = util.promisify<NewAddressRequest, Metadata, NewAddressResponse>(
-  bitcoinBridgeClient.newAddress.bind(bitcoinBridgeClient),
-)
-const findAddressByExternalId = util.promisify<
-  FindAddressByExternalIdRequest,
-  Metadata,
-  FindAddressByExternalIdResponse
->(bitcoinBridgeClient.findAddressByExternalId.bind(bitcoinBridgeClient))
-const getWalletBalanceSummary = util.promisify<
-  GetWalletBalanceSummaryRequest,
-  Metadata,
-  GetWalletBalanceSummaryResponse
->(bitcoinBridgeClient.getWalletBalanceSummary.bind(bitcoinBridgeClient))
-
-export const BriaPayloadType = {
-  UtxoDetected: "utxo_detected",
-  UtxoSettled: "utxo_settled",
-  PayoutSubmitted: "payout_submitted",
-  PayoutCommitted: "payout_committed",
-  PayoutBroadcast: "payout_broadcast",
-  PayoutSettled: "payout_settled",
-} as const
-
-const eventRepo = BriaEventRepo()
+const { streamBuilder, FibonacciBackoff } = GrpcStreamClient
 
 export const BriaSubscriber = () => {
   const metadata = new Metadata()
@@ -87,12 +42,9 @@ export const BriaSubscriber = () => {
 
   const subscribeToAll = async (
     eventHandler: BriaEventHandler,
-  ): Promise<ListenerWrapper | BriaEventError> => {
-    const subscribeAll = bitcoinBridgeClient.subscribeAll.bind(bitcoinBridgeClient)
-
-    let listenerWrapper: ListenerWrapper
+  ): Promise<Stream<RawBriaEvent> | BriaEventError> => {
     try {
-      const lastSequence = await eventRepo.getLatestSequence()
+      const lastSequence = await BriaEventRepo().getLatestSequence()
       if (lastSequence instanceof Error) {
         return lastSequence
       }
@@ -101,56 +53,27 @@ export const BriaSubscriber = () => {
       request.setAugment(true)
       request.setAfterSequence(lastSequence)
 
-      listenerWrapper = new ListenerWrapper(
-        subscribeAll(request, metadata),
-        (error: Error) => {
-          if (!error.message.includes("CANCELLED")) {
-            listenerWrapper._listener.cancel()
-            throw error
-          }
-        },
-      )
+      const onDataHandler = wrapAsyncToRunInSpan({
+        root: true,
+        namespace: "service.bria",
+        fnName: "subscribeToAllHandler",
+        fn: eventDataHandler(eventHandler),
+      })
+
+      return streamBuilder<RawBriaEvent, SubscribeAllRequest>(subscribeAll)
+        .withOptions({ retry: true, acceptDataOnReconnect: false })
+        .withRequest(request)
+        .withMetadata(metadata)
+        .onData(onDataHandler)
+        .onError((_, error) => baseLogger.info({ error }, "Error subscribeToAll stream"))
+        .onRetry((_, { detail }) =>
+          baseLogger.info({ ...detail }, "Retry subscribeToAll stream"),
+        )
+        .withBackoff(new FibonacciBackoff(30000, 7))
+        .build()
     } catch (error) {
       return new UnknownBriaEventError(error.message || error)
     }
-
-    listenerWrapper._setDataHandler((rawEvent: RawBriaEvent) => {
-      asyncRunInSpan(
-        "service.bria.eventReceived",
-        {
-          attributes: {
-            [SemanticAttributes.CODE_FUNCTION]: "eventReceived",
-            [SemanticAttributes.CODE_NAMESPACE]: "services.bria",
-            rawEvent: JSON.stringify(rawEvent.toObject()),
-          },
-        },
-        async () => {
-          const event = translate(rawEvent)
-          if (event instanceof Error) {
-            recordExceptionInCurrentSpan({ error: event })
-            throw event
-          }
-          const result = await eventHandler(event)
-
-          if (result instanceof Error) {
-            recordExceptionInCurrentSpan({ error: result })
-            const resubscribe = await subscribeToAll(eventHandler)
-            if (resubscribe instanceof Error) {
-              throw resubscribe
-            }
-            listenerWrapper._merge(resubscribe)
-          }
-
-          const res = await eventRepo.persistEvent(event)
-          if (res instanceof Error) {
-            recordExceptionInCurrentSpan({ error: res })
-            throw res
-          }
-        },
-      )
-    })
-
-    return listenerWrapper
   }
 
   return {
@@ -198,7 +121,7 @@ export const NewOnChainService = (): INewOnChainService => {
       const response = await newAddress(request, metadata)
       return { address: response.getAddress() as OnChainAddress }
     } catch (err) {
-      if (err.code == status.ALREADY_EXISTS) {
+      if (err.code === status.ALREADY_EXISTS) {
         return new OnChainAddressAlreadyCreatedForRequestIdError()
       }
       const errMsg = typeof err === "string" ? err : err.message
@@ -239,140 +162,12 @@ export const NewOnChainService = (): INewOnChainService => {
   })
 }
 
-const translate = (rawEvent: RawBriaEvent): BriaEvent | BriaEventError => {
-  const sequence = rawEvent.getSequence()
-  const rawAugmentation = rawEvent.getAugmentation()
-
-  if (!rawAugmentation) {
-    return new EventAugmentationMissingError()
-  }
-  let augmentation: BriaEventAugmentation | undefined = undefined
-  const rawInfo = rawAugmentation.getAddressInfo()
-  if (rawInfo) {
-    const info = rawInfo.toObject()
-    augmentation = {
-      addressInfo: {
-        address: info.address as OnChainAddress,
-        externalId: info.externalId,
-      },
-    }
-  }
-  if (augmentation === undefined) {
-    return new ExpectedAddressInfoMissingInEventError()
-  }
-
-  let payload: BriaPayload | undefined
-  let rawPayload
-  switch (rawEvent.getPayloadCase()) {
-    case RawBriaEvent.PayloadCase.PAYLOAD_NOT_SET:
-      return new NoPayloadFoundError()
-    case RawBriaEvent.PayloadCase.UTXO_DETECTED:
-      rawPayload = rawEvent.getUtxoDetected()
-      if (rawPayload === undefined) {
-        return new ExpectedUtxoDetectedPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.UtxoDetected,
-        txId: rawPayload.getTxId() as OnChainTxHash,
-        vout: rawPayload.getVout() as OnChainTxVout,
-        address: rawPayload.getAddress() as OnChainAddress,
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-      }
-      break
-    case RawBriaEvent.PayloadCase.UTXO_SETTLED:
-      rawPayload = rawEvent.getUtxoSettled()
-      if (rawPayload === undefined) {
-        return new ExpectedUtxoSettledPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.UtxoSettled,
-        txId: rawPayload.getTxId() as OnChainTxHash,
-        vout: rawPayload.getVout() as OnChainTxVout,
-        address: rawPayload.getAddress() as OnChainAddress,
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-        blockNumber: rawPayload.getBlockHeight(),
-      }
-      break
-    case RawBriaEvent.PayloadCase.PAYOUT_SUBMITTED:
-      rawPayload = rawEvent.getPayoutSubmitted()
-      if (rawPayload === undefined) {
-        return new ExpectedPayoutSubmittedPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.PayoutSubmitted,
-        id: rawPayload.getId(),
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-      }
-      break
-    case RawBriaEvent.PayloadCase.PAYOUT_COMMITTED:
-      rawPayload = rawEvent.getPayoutCommitted()
-      if (rawPayload === undefined) {
-        return new ExpectedPayoutCommittedPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.PayoutCommitted,
-        id: rawPayload.getId(),
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-      }
-      break
-    case RawBriaEvent.PayloadCase.PAYOUT_BROADCAST:
-      rawPayload = rawEvent.getPayoutBroadcast()
-      if (rawPayload === undefined) {
-        return new ExpectedPayoutBroadcastPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.PayoutBroadcast,
-        id: rawPayload.getId(),
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-      }
-      break
-    case RawBriaEvent.PayloadCase.PAYOUT_SETTLED:
-      rawPayload = rawEvent.getPayoutSettled()
-      if (rawPayload === undefined) {
-        return new ExpectedPayoutSettledPayloadNotFoundError()
-      }
-      payload = {
-        type: BriaPayloadType.PayoutSettled,
-        id: rawPayload.getId(),
-        satoshis: {
-          amount: BigInt(rawPayload.getSatoshis()),
-          currency: WalletCurrency.Btc,
-        },
-      }
-      break
-    default:
-      return new UnknownPayloadTypeReceivedError()
-  }
-
-  return {
-    payload,
-    augmentation,
-    sequence,
-  }
-}
-
 export const KnownBriaErrorDetails = {
   DuplicateRequestIdAddressCreate:
     /duplicate key value violates unique constraint.*bria_addresses_account_id_external_id_key/,
 } as const
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const constructMetadata = (metadataObj: { [key: string]: any }): Struct => {
+const constructMetadata = (metadataObj: JSONObject): Struct => {
   const metadata = new Struct()
 
   for (const key in metadataObj) {
@@ -394,7 +189,7 @@ const constructMetadata = (metadataObj: { [key: string]: any }): Struct => {
           new ListValue().setValuesList(
             fieldValue.map((item) => {
               const listItemValue = new Value()
-              listItemValue.setStringValue(item)
+              listItemValue.setStringValue(`${item}`)
               return listItemValue
             }),
           ),
@@ -404,7 +199,7 @@ const constructMetadata = (metadataObj: { [key: string]: any }): Struct => {
           value.setListValue(fetchedListValue)
         }
       } else if (typeof fieldValue === "object") {
-        value.setStructValue(constructMetadata(fieldValue))
+        value.setStructValue(constructMetadata(fieldValue as JSONObject))
       }
 
       metadata.getFieldsMap().set(key, value)
